+++ packages/cli/kit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @damatjs/cli-kit
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @damatjs/installer@1.0.5
+  - @damatjs/cli-support@1.0.5
+  - @damatjs/cli@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes