+++ packages/core/cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @damatjs/cli
 
+## 1.0.5
+
 ## 1.0.4
 
 ### Patch Changes