+++ packages/core/events/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @damatjs/events
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+- @damatjs/durability@1.0.5
+- @damatjs/logger@1.0.5
+- @damatjs/redis@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes