+++ packages/core/redis/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+- @damatjs/logger@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes