+++ packages/core/schema-codegen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @damatjs/schema-codegen
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/orm-type@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes