+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @damatjs/framework
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @damatjs/installer@1.0.5
+  - @damatjs/provider-auth@1.0.5
+  - @damatjs/deps@1.0.5
+  - @damatjs/durability@1.0.5
+  - @damatjs/events@1.0.5
+  - @damatjs/jobs@1.0.5
+  - @damatjs/link@1.0.5
+  - @damatjs/logger@1.0.5
+  - @damatjs/orm-connector@1.0.5
+  - @damatjs/orm-type@1.0.5
+  - @damatjs/pipelines@1.0.5
+  - @damatjs/provider@1.0.5
+  - @damatjs/redis@1.0.5
+  - @damatjs/services@1.0.5
+  - @damatjs/types@1.0.5
+  - @damatjs/workflow-engine@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes