+++ packages/module-generator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @damatjs/module-generator
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/logger@1.0.5
+- @damatjs/orm-migration@1.0.5
+- @damatjs/orm-model@1.0.5
+- @damatjs/orm-type@1.0.5
+- @damatjs/schema-codegen@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes