+++ packages/orm/processor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @damatjs/orm-processor
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+- @damatjs/orm-model@1.0.5
+- @damatjs/orm-type@1.0.5
+- @damatjs/types@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes