+++ packages/orm/type/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @damatjs/orm-type
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes