+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @damatjs/provider
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+- @damatjs/services@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes