+++ packages/workflow-engine/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @damatjs/workflow-engine
 
+## 1.0.5
+
+### Patch Changes
+
+- @damatjs/deps@1.0.5
+- @damatjs/logger@1.0.5
+- @damatjs/redis@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes